requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
ultralytics==8.0.0
opencv-python-headless==4.5.5.64
pillow==10.0.1
python-multipart==0.0.6
numpy==1.24.3